requirements.txt
openai
langchain
langchain_openai
praw
python-dotenv
